pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
